SEQNG-335 Move operations to a separate object
